API

Public API

ONet.ONet — Module.

ONet

Application to upload the ONet data to the database.

source

ONet.onet! — Method.

onet!()

Runs the application to refresh all the data in the database.

source

ONet.sdad_setup! — Method.

sdad_setup!(;db_user::AbstractString = "",
             db_pwd::AbstractString = "",
             token::AbstractString = "")

This function writes the configurations to confs/config.simple

Arguments

db_user: Your username for the postgis_2 database (i.e., your UVA computing ID).

db_pwd: The password for the postgis_2 database.

onet_auth: Basic authentication token for the O*Net REST API.

Examples

Good!

julia> sdad_setup!(db_user = "jbs3hp",
                   db_pwd = "MyVerySafePwd",
                   onet_auth = "ASZ1QqATfYXdOLFnr6EVfxyHymx4kTGMLQgXCs==")
😃

Incomplete!

julia> sdad_setup!(db_user = "jbs3hp",
                   db_pwd = "MyVerySafePwd")
Warning: github_login has not been defined.
😞

Updating the configuration file.

julia> using ONet
julia> sdad_setup!(db_user = "jbs3hp",
                   db_pwd = "MyVerySafePwd",
                   onet_auth = "ASZ1QqATfYXdOLFnr6EVfxyHymx4kTGMLQgXCs==")
julia> exit()
> julia
julia> using ONet # config up-to-date

Note

Updating the configuration file requires a restart of the session.

source

Private API

ONet.db_host — Constant.

db_host = "postgis_2"

Host for the database.

source

ONet.db_port — Constant.

db_port = 5432

Port for the postgis_2 in the database.

source

ONet.db_pwd — Constant.

db_pwd::String

The password for the database.

source

ONet.db_user — Constant.

db_user::String

The username for the database (i.e., your UVA computing ID).

source

ONet.dbname — Constant.

dbname = "sdad_data"

Database for the SDAD data.

source

ONet.headers — Constant.

headers = ["Authorization" => "Basic $onet_auth",
           "Accept" => "application/json"]

Headers to be passed to the ONet REST API.

source

ONet.onet_auth — Constant.

onet_auth::String

The application O*Net basic authentication token.

source

ONet.createtable! — Method.

createtable!(name)

Creates a table in the database based on the O*NET API specification.

source

ONet.dbconnect — Method.

dbconnect()::Connection

Returns a connection to the burningglass database in postgis2.

source

ONet.drop_duplicates! — Method.

drop_duplicates!(name::AbstractString)

Drop duplicate records from the table.

source

ONet.isdone — Method.

isdone(name::AbstractString)::Bool

Are all available records for that table in the database?

source

ONet.onettable — Method.

onettable(table)

Given a table node, it returns a row with the table information.

source

ONet.onettables! — Method.

onettables!()

Uploads a _metadata table to the database.

source

ONet.processtable! — Method.

processtable!(name::AbstractString)

Adds the table data to the database.

source

ONet.varschema — Method.

varschema(variable)::String

Returns a suitable schema for the database declaration from the table information.

source